requirements.txt
scapy>=2.5.0
numpy>=1.24.0
pandas>=2.0.0
tensorflow>=2.12.0
scikit-learn>=1.3.0
lime>=0.2.0.1
shap>=0.42.0
joblib>=1.3.0
mac-vendor-lookup>=0.1.12
python-dotenv
requests
plyer
